@@ -8,7 +8,7 @@ import { AxiosResponseHeaders, RawAxiosResponseHeaders } from 'axios';
 import { randomUUID } from 'crypto';
 import debugBuilder from 'debug';
 import { EventEmitter } from 'events';
-import getStream = require('get-stream');
+import getStream from 'get-stream';
 import * as _ from 'lodash';
 import { HttpClient } from './http';
 import { IHeaders, IHttpClient, IMTOMAttachments, IOptions, ISecurity, SoapMethod, SoapMethodAsync } from './types';
@@ -25,8 +25,7 @@ export interface ISoapError extends Error {
   body?;
 }
 
-// tslint:disable unified-signatures
-// tslint:disable-next-line:interface-name
+//eslint-disable-next-line @typescript-eslint/no-unsafe-declaration-merging
 export interface Client {
   emit(event: 'request', xml: string, eid: string): boolean;
   emit(event: 'message', message: string, eid: string): boolean;
@@ -43,6 +42,7 @@ export interface Client {
   on(event: 'response', listener: (body: any, response: any, eid: string) => void): this;
 }
 
+//eslint-disable-next-line @typescript-eslint/no-unsafe-declaration-merging
 export class Client extends EventEmitter {
   [method: string]: any;
   /** contains last full soap request for client logging */
@@ -119,13 +119,13 @@ export class Client extends EventEmitter {
     this.httpHeaders = null;
   }
 
-  public addBodyAttribute(bodyAttribute: any, name?: string, namespace?: string, xmlns?: string): void {
+  public addBodyAttribute(bodyAttribute: any): void {
     if (!this.bodyAttributes) {
       this.bodyAttributes = [];
     }
     if (typeof bodyAttribute === 'object') {
       let composition = '';
-      Object.getOwnPropertyNames(bodyAttribute).forEach((prop, idx, array) => {
+      Object.getOwnPropertyNames(bodyAttribute).forEach((prop) => {
         composition += ' ' + prop + '="' + bodyAttribute[prop] + '"';
       });
       bodyAttribute = composition;
@@ -280,12 +280,11 @@ export class Client extends EventEmitter {
     switch (typeof soapHeader) {
       case 'object':
         return this.wsdl.objectToXML(soapHeader, name, namespace, xmlns, true);
-      case 'function':
+      case 'function': {
+        //eslint-disable-next-line @typescript-eslint/no-this-alias
         const _this = this;
-        // arrow function does not support arguments variable
-        // tslint:disable-next-line
-        return function () {
-          const result = soapHeader.apply(null, arguments);
+        return (...args: any) => {
+          const result = soapHeader.apply(null, [...args]);
 
           if (typeof result === 'object') {
             return _this.wsdl.objectToXML(result, name, namespace, xmlns, true);
@@ -293,6 +292,7 @@ export class Client extends EventEmitter {
             return result;
           }
         };
+      }
       default:
         return soapHeader;
     }
@@ -353,7 +353,7 @@ export class Client extends EventEmitter {
       }
       if (!result) {
         ['Response', 'Out', 'Output'].forEach((term) => {
-          if (obj.Body.hasOwnProperty(name + term)) {
+          if (Object.hasOwnProperty.call(obj.Body, name + term)) {
             return result = obj.Body[name + term];
           }
         });
@@ -413,7 +413,7 @@ export class Client extends EventEmitter {
     if ((style === 'rpc') && ((input.parts || input.name === 'element') || args === null)) {
       assert.ok(!style || style === 'rpc', 'invalid message definition for document style binding');
       message = this.wsdl.objectToRpcXML(name, args, alias, ns, (input.name !== 'element'));
-      (method.inputSoap === 'encoded') && (encoding = 'soap:encodingStyle="http://schemas.xmlsoap.org/soap/encoding/" ');
+      if (method.inputSoap === 'encoded') encoding = 'soap:encodingStyle="http://schemas.xmlsoap.org/soap/encoding/" ';
     } else {
       assert.ok(!style || style === 'document', 'invalid message definition for rpc style binding');
       // pass `input.$lookupType` if `input.$type` could not be found
@@ -482,7 +482,7 @@ export class Client extends EventEmitter {
     const tryJSONparse = (body) => {
       try {
         return JSON.parse(body);
-      } catch (err) { return undefined; }
+      } catch { return undefined; }
     };
 
     if (this.streamAllowed && typeof this.httpClient.requestStream === 'function') {
@@ -497,7 +497,7 @@ export class Client extends EventEmitter {
         if (this.returnSaxStream || !err.response || !err.response.data) {
           callback(err, undefined, undefined, undefined, xml);
         } else {
-          err.response.data.on('close', (e) => {
+          err.response.data.on('close', () => {
             callback(err, undefined, undefined, undefined, xml);
           });
           err.response.data.on('data', (e) => {
