@@ -7,15 +7,9 @@ import { EventEmitter } from 'events';
 import * as http from 'http';
 import * as url from 'url';
 import { IOneWayOptions, IServerOptions, IServices, ISoapFault, ISoapServiceMethod } from './types';
-import { findPrefix } from './utils';
 import { WSDL } from './wsdl';
 import { BindingElement, IPort } from './wsdl/elements';
-
-let zlib;
-try {
-  zlib = require('zlib');
-} catch (error) {
-}
+import zlib from "zlib"
 
 interface IExpressApp {
   route;
@@ -46,8 +40,7 @@ function getDateString(d) {
     + pad(d.getUTCSeconds()) + 'Z';
 }
 
-// tslint:disable unified-signatures
-// tslint:disable-next-line:interface-name
+//eslint-disable-next-line  @typescript-eslint/no-unsafe-declaration-merging
 export interface Server {
   emit(event: 'request', request: any, methodName: string): boolean;
   emit(event: 'headers', headers: any, methodName: string): boolean;
@@ -71,6 +64,7 @@ interface IExecuteMethodOptions {
   style?: 'document' | 'rpc';
 }
 
+//eslint-disable-next-line  @typescript-eslint/no-unsafe-declaration-merging
 export class Server extends EventEmitter {
   public path: string | RegExp;
   public services: IServices;
@@ -177,12 +171,11 @@ export class Server extends EventEmitter {
     switch (typeof soapHeader) {
       case 'object':
         return this.wsdl.objectToXML(soapHeader, name, namespace, xmlns, true);
-      case 'function':
+      case 'function': {
+        //eslint-disable-next-line @typescript-eslint/no-this-alias
         const _this = this;
-        // arrow function does not support arguments variable
-        // tslint:disable-next-line
-        return function () {
-          const result = soapHeader.apply(null, arguments);
+        return (...args: any) => {
+          const result = soapHeader.apply(null, [...args]);
 
           if (typeof result === 'object') {
             return _this.wsdl.objectToXML(result, name, namespace, xmlns, true);
@@ -190,6 +183,7 @@ export class Server extends EventEmitter {
             return result;
           }
         };
+      }
       default:
         return soapHeader;
     }
@@ -233,7 +227,6 @@ export class Server extends EventEmitter {
 
   private _requestListener(req: Request, res: Response) {
     const reqParse = url.parse(req.url);
-    const reqPath = reqParse.pathname;
     const reqQuery = reqParse.search;
 
     if (typeof this.log === 'function') {
@@ -285,12 +278,12 @@ export class Server extends EventEmitter {
 
   private _getSoapAction(req: Request) {
     if (typeof req.headers.soapaction === 'undefined') {
-         return;
-     }
+      return;
+    }
     const soapAction: string = req.headers.soapaction as string;
     return (soapAction.indexOf('"') === 0)
-         ? soapAction.slice(1, -1)
-         : soapAction;
+      ? soapAction.slice(1, -1)
+      : soapAction;
   }
 
   private _process(input, req: Request, res: Response, cb: (result: any, statusCode?: number) => any) {
@@ -512,7 +505,7 @@ export class Server extends EventEmitter {
 
     try {
       method = this.services[serviceName][portName][methodName];
-    } catch (error) {
+    } catch {
       return callback(this._envelope('', headers, includeTimestamp));
     }
 
@@ -595,10 +588,7 @@ export class Server extends EventEmitter {
   }
 
   private _envelope(body, headers, includeTimestamp) {
-    const defs = this.wsdl.definitions;
-    const ns = defs.$targetNamespace;
     const encoding = '';
-    const alias = findPrefix(defs.xmlns, ns);
 
     const envelopeDefinition = this.wsdl.options.forceSoap12Headers
       ? 'http://www.w3.org/2003/05/soap-envelope'
