@@ -13,7 +13,7 @@ interface INamespace {
  */
 class NamespaceScope {
   public parent: NamespaceScope;
-  public namespaces: {[key: string]: INamespace};
+  public namespaces: { [key: string]: INamespace };
 
   constructor(parent: NamespaceScope) {
     this.parent = parent;
@@ -32,7 +32,7 @@ class NamespaceScope {
         return 'http://www.w3.org/XML/1998/namespace';
       case 'xmlns':
         return 'http://www.w3.org/2000/xmlns/';
-      default:
+      default: {
         const nsUri = this.namespaces[prefix];
         /*jshint -W116 */
         if (nsUri != null) {
@@ -42,6 +42,7 @@ class NamespaceScope {
         } else {
           return null;
         }
+      }
     }
   }
 
@@ -59,7 +60,7 @@ class NamespaceScope {
           prefix: 'xmlns',
           declared: true,
         };
-      default:
+      default: {
         const mapping = this.namespaces[prefix];
         /*jshint -W116 */
         if (mapping != null) {
@@ -69,6 +70,7 @@ class NamespaceScope {
         } else {
           return null;
         }
+      }
     }
   }
 
