@@ -114,8 +114,8 @@ export class Element {
   }
 
   public deleteFixedAttrs() {
-    this.children && this.children.length === 0 && delete this.children;
-    this.xmlns && Object.keys(this.xmlns).length === 0 && delete this.xmlns;
+    if (this.children && this.children.length === 0) delete this.children;
+    if (this.xmlns && Object.keys(this.xmlns).length === 0) delete this.xmlns;
     delete this.nsName;
     delete this.prefix;
     delete this.name;
@@ -160,6 +160,7 @@ export class Element {
     }
   }
 
+  //eslint-disable-next-line @typescript-eslint/no-unused-vars
   public addChild(child: Element) {
     return;
   }
@@ -168,6 +169,7 @@ export class Element {
     throw new Error('Found unexpected element (' + name + ') inside ' + this.nsName);
   }
 
+  //eslint-disable-next-line @typescript-eslint/no-unused-vars
   public description(definitions?: DefinitionsElement, xmlns?: IXmlNs): any {
     return this.$name || this.name;
   }
@@ -219,17 +221,12 @@ export class ElementElement extends Element {
 
     // Check minOccurs / maxOccurs attributes to see if this element is a list
     // These are default values for an element
-    let minOccurs = 1;
     let maxOccurs = 1;
 
     if (this.$maxOccurs === 'unbounded') {
       maxOccurs = Infinity;
-    } else if (Boolean(this.$maxOccurs)) {
+    } else if (this.$maxOccurs) {
       maxOccurs = parseInt(this.$maxOccurs, 10);
-    }
-
-    if (Boolean(this.$minOccurs)) {
-      minOccurs = parseInt(this.$minOccurs, 10);
     }
 
     const isMany = maxOccurs > 1;
@@ -385,7 +382,7 @@ export class SimpleTypeElement extends Element {
     'restriction',
   ]);
 
-  public description(definitions: DefinitionsElement) {
+  public description() {
     for (const child of this.children) {
       if (child instanceof RestrictionElement) {
         return [this.$name, child.description()].filter(Boolean).join('|');
@@ -616,6 +613,7 @@ export class AttributeElement extends Element {
   public $type?: string;
   public $use?: string;
 
+  //eslint-disable-next-line @typescript-eslint/no-unused-vars
   public description(definitions: DefinitionsElement) {
     return {
       type: this.$type,
@@ -741,7 +739,8 @@ export class MessageElement extends Element {
       // rpc encoding
       this.parts = {};
       delete this.element;
-      for (let i = 0; part = this.children[i]; i++) {
+      for (let i = 0; i < this.children.length; i++) {
+        const part = this.children[i] as any;
         if (part.name === 'documentation') {
           // <wsdl:documentation can be present under <wsdl:message>
           continue;
@@ -816,7 +815,7 @@ export class MessageElement extends Element {
     let resolvedType = '^';
     const excluded = this.ignoredNamespaces.concat('xs'); // do not process $type values wich start with
 
-    if (element.hasOwnProperty('$type') && typeof element.$type === 'string') {
+    if (Object.hasOwnProperty.call(element, '$type') && typeof element.$type === 'string') {
       if (excluded.indexOf(element.$type.split(':')[0]) === -1) {
         resolvedType += ('_' + element.$type + '#' + element.$name);
       }
@@ -869,7 +868,7 @@ export class SchemaElement extends Element {
 
     // Merge attributes from source without overwriting our's
     _.merge(this, _.pickBy(source, (value, key) => {
-      return key.startsWith('$') && !this.hasOwnProperty(key);
+      return key.startsWith('$') && !Object.hasOwnProperty.call(this, key);
     }));
 
     return this;
@@ -911,17 +910,17 @@ export class TypesElement extends Element {
     assert(child instanceof SchemaElement);
 
     const childInclude = child.includes.find((e: object) => {
-      return e.hasOwnProperty('namespace');
+      return Object.hasOwnProperty.call(e, 'namespace');
     });
     const childIncludeNs: string = (
       (typeof childInclude !== 'undefined' &&
-        childInclude.hasOwnProperty('namespace') &&
+        Object.hasOwnProperty.call(childInclude, 'namespace') &&
         child instanceof SchemaElement) ?
         childInclude.namespace :
         undefined);
     const targetNamespace = child.$targetNamespace || child.includes[0]?.namespace || childIncludeNs;
 
-    if (!this.schemas.hasOwnProperty(targetNamespace)) {
+    if (!Object.hasOwnProperty.call(this.schemas, targetNamespace)) {
       this.schemas[targetNamespace] = child;
     } else {
       if (targetNamespace === child.$targetNamespace) {
@@ -960,7 +959,8 @@ export class OperationElement extends Element {
 
   public postProcess(definitions: DefinitionsElement, tag: string) {
     const children = this.children;
-    for (let i = 0, child; child = children[i]; i++) {
+    for (let i = 0; i < children.length; i++) {
+      const child = children[i] as any
       if (child.name !== 'input' && child.name !== 'output') {
         continue;
       }
@@ -1009,7 +1009,8 @@ export class PortTypeElement extends Element {
     if (typeof children === 'undefined') {
       return;
     }
-    for (let i = 0, child; child = children[i]; i++) {
+    for (let i = 0; i < children.length; i++) {
+      const child = children[i] as any
       if (child.name !== 'operation') {
         continue;
       }
@@ -1070,13 +1071,14 @@ export class BindingElement extends Element {
       portType.postProcess(definitions);
       this.methods = portType.methods;
 
-      for (let i = 0, child; child = children[i]; i++) {
+      for (let i = 0; i < children.length; i++) {
+        const child = children[i] as any
         if (child.name !== 'operation') {
           continue;
         }
         child.postProcess(definitions, 'binding');
         children.splice(i--, 1);
-        child.style || (child.style = style);
+        if (!child.style) child.style = style;
         const method = this.methods[child.$name];
 
         if (method) {
@@ -1084,8 +1086,8 @@ export class BindingElement extends Element {
           method.soapAction = child.soapAction;
           method.inputSoap = child.input || null;
           method.outputSoap = child.output || null;
-          method.inputSoap && method.inputSoap.deleteFixedAttrs();
-          method.outputSoap && method.outputSoap.deleteFixedAttrs();
+          if (method.inputSoap) method.inputSoap.deleteFixedAttrs();
+          if (method.outputSoap) method.outputSoap.deleteFixedAttrs();
         }
       }
     }
@@ -1134,7 +1136,8 @@ export class ServiceElement extends Element {
     const children = this.children;
     const bindings = definitions.bindings;
     if (children && children.length > 0) {
-      for (let i = 0, child; child = children[i]; i++) {
+      for (let i = 0; i < children.length; i++) {
+        const child = children[i] as any;
         if (child.name !== 'port') {
           continue;
         }
@@ -1216,7 +1219,6 @@ export class DefinitionsElement extends Element {
       }
     } else if (child instanceof ServiceElement) {
       this.services[child.$name] = child;
-    } else if (child instanceof DocumentationElement) {
     }
     this.children.pop();
   }
