@@ -48,15 +48,20 @@ describe('SOAP Client', function () {
 
   it('lastElapsedTime is computed', function (done) {
     soap.createClientAsync(__dirname + '/wsdl/hello.wsdl', { endpoint: `${url}${servicePath}` }).then(function (client) {
-      assert.ok(client);
-      console.log('client created:', client.describe());
-      client.sayHelloAsync({ firstName: 'LastElapsedTime Tester' }, { time: true }).then(() => {
-        assert.ok(Object.prototype.hasOwnProperty.call(client, 'lastElapsedTime'));
-        assert.ok(typeof client.lastElapsedTime === 'number');
-        console.log('api finished in ms:', client.lastElapsedTime);
-        assert.ok(client.lastElapsedTime > 0);
-        done();
-      });
+      try {
+        assert.ok(client);
+        console.log('client created:', client.describe());
+        client.sayHelloAsync({ firstName: 'LastElapsedTime Tester' }, { time: true }).then(() => {
+          assert.ok(Object.prototype.hasOwnProperty.call(client, 'lastElapsedTime'));
+          assert.ok(typeof client.lastElapsedTime === 'number');
+          console.log('api finished in ms:', client.lastElapsedTime);
+          assert.ok(client.lastElapsedTime > 0);
+        });
+      } catch (err) {
+        done(err);
+        throw err;
+      }
+      done();
     });
   });
 
