@@ -5,7 +5,7 @@ var http = require('http');
 var soap = require('../lib/soap');
 var server;
 
-describe('Preserve data encoding from endpoint response', function() {
+describe('Preserve data encoding from endpoint response', function () {
   var wsdl = __dirname + '/wsdl/hello.wsdl';
   var expectedString = "àáÁÉÈÀçãü";
   var xml = `<?xml version=\"1.0\" encoding=\"iso-8859-1\"?><soap:Envelope xmlns:soap=\"http://schemas.xmlsoap.org/soap/envelope/\"  xmlns:tns=\"http://www.examples.com/wsdl/HelloService.wsdl\"><soap:Body><tns:sayHelloResponse>${expectedString}</tns:sayHelloResponse></soap:Body></soap:Envelope>`;
@@ -14,19 +14,19 @@ describe('Preserve data encoding from endpoint response', function() {
     "binary"
   );
 
-  before(function(done) {
-    server = http.createServer(function(req, res) {
+  before(function (done) {
+    server = http.createServer(function (req, res) {
       res.statusCode = 200;
       res.end(xmlEncoded);
     }).listen(51515, done);
   });
 
-  after(function() {
+  after(function () {
     server.close();
   });
 
   it('Should read special characters with enconding option with success',
-    function(done) {
+    function (done) {
       var url = 'http://' + server.address().address + ':' + server.address().port;
 
       if (server.address().address === '0.0.0.0' || server.address().address === '::') {
@@ -38,10 +38,10 @@ describe('Preserve data encoding from endpoint response', function() {
         {
           endpoint: url,
           disableCache: true, // disable wsdl cache, otherwise 'mocha test/client-response-options-test.js test/response-preserve-whitespace-test.js' will fail.
-          parseReponseAttachments:true,
-          encoding:'latin1',
+          parseReponseAttachments: true,
+          encoding: 'latin1',
         },
-        function(err, client) {
+        function (err, client) {
           if (err) {
             console.log(err);
             throw err;
@@ -51,12 +51,12 @@ describe('Preserve data encoding from endpoint response', function() {
             {
               firstName: 'hello world'
             },
-            function(err, result, rawResponse, soapHeader, rawRequest) {
+            function (err, result, rawResponse, soapHeader, rawRequest) {
               if (err) {
                 console.log(err);
                 throw err;
               }
-              
+
               assert.strictEqual(expectedString, result);
               done();
             }
@@ -66,41 +66,41 @@ describe('Preserve data encoding from endpoint response', function() {
     });
 
   it('Should read special characters with enconding option with error',
-  function(done) {
-    var url = 'http://' + server.address().address + ':' + server.address().port;
+    function (done) {
+      var url = 'http://' + server.address().address + ':' + server.address().port;
 
-    if (server.address().address === '0.0.0.0' || server.address().address === '::') {
-      url = 'http://127.0.0.1:' + server.address().port;
-    }
-
-    soap.createClient(
-      wsdl,
-      {
-        endpoint: url,
-        disableCache: true, // disable wsdl cache, otherwise 'mocha test/client-response-options-test.js test/response-preserve-whitespace-test.js' will fail.
-        parseReponseAttachments:true,
-      },
-      function(err, client) {
-        if (err) {
-          console.log(err);
-          throw err;
-        }
-
-        client.sayHello(
-          {
-            firstName: 'hello world'
-          },
-          function(err, result, rawResponse, soapHeader, rawRequest) {
-            if (err) {
-              console.log(err);
-              throw err;
-            }
-            assert.strictEqual('���������', result);
-            done();
-          }
-        );
+      if (server.address().address === '0.0.0.0' || server.address().address === '::') {
+        url = 'http://127.0.0.1:' + server.address().port;
       }
-    );
-  });
+
+      soap.createClient(
+        wsdl,
+        {
+          endpoint: url,
+          disableCache: true, // disable wsdl cache, otherwise 'mocha test/client-response-options-test.js test/response-preserve-whitespace-test.js' will fail.
+          parseReponseAttachments: true,
+        },
+        function (err, client) {
+          if (err) {
+            console.log(err);
+            throw err;
+          }
+
+          client.sayHello(
+            {
+              firstName: 'hello world'
+            },
+            function (err, result, rawResponse, soapHeader, rawRequest) {
+              if (err) {
+                console.log(err);
+                throw err;
+              }
+              assert.strictEqual('���������', result);
+              done();
+            }
+          );
+        }
+      );
+    });
 
 });
