@@ -15,7 +15,7 @@ describe('PasswordDigest', function () {
 	});
 
 	it('should calculate a valid passworddigest ', function () {
-		
+
 		var result = Utils.passwordDigest(nonce, created, password);
 		assert.equal(result, expected);
 	});
